+++ lib/investigationCollection/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./dnsProvider";
 export * from "./orchestrator";
 export * from "./capabilityRegistry";
+export * from "./publicWebProvider";